refactor: create directory for screens
