+++ src/main/kotlin/view/screens/MainScreen.kt
@@ -1,4 +1,4 @@
-package view
+package view.screens
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
